setup(ProductImportRow > ImportedProduct): Nested UI components added to deal with each row from imported product during onboard
