+++ Interfaces/ForTenantsInterface/components/PlanIdInterface/components/ProductsStep/components/AddProductsInterface/components/ProductImportPreviewTable/components/ImportedProduct/components/PricingRow/index.ts
@@ -1,0 +1,1 @@
+export { default as PricingRow } from "./PricingRow";
