+++ Interfaces/ForTenantsInterface/components/PlanIdInterface/components/ProductsStep/components/AddProductsInterface/components/ProductImportPreviewTable/components/ImportedProduct/index.ts
@@ -1,0 +1,1 @@
+export { default as ImportedProduct } from "./ImportedProduct";
