--- Interfaces/ForTenantsInterface/components/PlanIdInterface/components/ProductsStep/components/AddProductsInterface/components/ProductImportPreviewTable/components/ProductImportRow/index.ts
@@ -1,1 +1,0 @@
-export { default as ProductImportRow } from "./ProductImportRow";
